Move tests/common into common/test
